final order + invoice + invoice details

+++ src/main/java/com/java04/wibucian/services/InvoiceService.java
@@ -2,10 +2,7 @@ package com.java04.wibucian.services;
 
 import com.java04.wibucian.dtos.InvoiceDTO;
 import com.java04.wibucian.models.*;
-import com.java04.wibucian.repositories.EmployeeRepository;
-import com.java04.wibucian.repositories.GroupTableRepository;
-import com.java04.wibucian.repositories.InvoiceNoMapPingRepository;
-import com.java04.wibucian.repositories.InvoiceRepository;
+import com.java04.wibucian.repositories.*;
 import com.java04.wibucian.vos.InvoiceQueryVO;
 import com.java04.wibucian.vos.InvoiceUpdateVO;
 import com.java04.wibucian.vos.InvoiceVO;
@@ -14,9 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.*;
 
 @Service
 public class InvoiceService {
@@ -24,6 +19,9 @@ public class InvoiceService {
 
     @Autowired
     private GroupTableRepository groupTableRepository;
+
+    @Autowired
+    private DetailInvoiceRepository detailInvoiceRepository;
 
     @Autowired
     private EmployeeRepository employeeRepository;
@@ -34,10 +32,15 @@ public class InvoiceService {
     @Autowired
     private InvoiceNoMapPingRepository invoiceNoMapPingRepository;
 
+    @Autowired ProductRepository productRepository;
+
     public List<InvoiceNoMapPing> getAllInvoiceNoMapPing() {
         return invoiceNoMapPingRepository.findAll();
     }
 
+    public List<Product> findAllProduct(){
+        return productRepository.findAll();
+    }
     public List<Invoice> findAll() {
         return invoiceRepository.findAll();
     }
@@ -53,8 +56,8 @@ public class InvoiceService {
         BeanUtils.copyProperties(vO, bean);
         bean.setGroupTable(groupTable);
         bean.setEmployee(employee);
+        bean.setTotalMoney(0.0);
         bean = invoiceRepository.save(bean);
-        System.out.println(bean);
         InvoiceNoMapPing invoiceNoMapPing = findByIdCustom(bean.getId());
         return invoiceNoMapPing.getIdInvoice();
     }
@@ -63,24 +66,65 @@ public class InvoiceService {
         String idGroupTabel = vO.getIdGroupTable();
         List<InvoiceNoMapPing> invoiceNoMapPings = findByIdGroupTableAndStatus(idGroupTabel, "1");
         if (invoiceNoMapPings != null) {
-
             InvoiceNoMapPing invoiceNoMapPing = invoiceNoMapPings.get(0);
             return invoiceNoMapPing.getIdInvoice();
-        }
-        else {
+        } else {
             vO.setStatus(1);
             return save(vO);
         }
     }
 
-    public void delete(String id) {
-        invoiceRepository.deleteById(id);
+    public HashMap<String, Object> delete(String id) {
+        HashMap<String, Object> map = new HashMap<>();
+        Invoice invoice = invoiceRepository.findById(id).get();
+        if (invoice != null) {
+            if (invoice.getStatus() != 2) {
+                Set<DetailInvoice> detailInvoices = invoice.getDetailInvoices();
+                if (detailInvoices != null) {
+                    detailInvoices.forEach(detailInvoice -> {
+                        System.out.println(detailInvoice.getId());
+                        detailInvoiceRepository.deleteById(detailInvoice.getId());
+                    });
+                }
+                map.put("check", true);
+                map.put("value", "Xóa thành công");
+                invoiceRepository.deleteById(id);
+
+            } else {
+                map.put("check", false);
+                map.put("value", "Đơn hàng đã thanh toán không thể xóa");
+
+            }
+        } else {
+            map.put("check", false);
+            map.put("value", "Không tồn tại đơn hàng này");
+        }
+        return map;
     }
 
     public void update(String id, InvoiceUpdateVO vO) {
         Invoice bean = requireOne(id);
-        BeanUtils.copyProperties(vO, bean);
-        invoiceRepository.save(bean);
+        if (bean != null) {
+            //check validate not null
+            if (vO.getStatus() != null) {
+                bean.setStatus(vO.getStatus());
+            }
+            if (vO.getToltalMoney() != null) {
+                bean.setTotalMoney(vO.getToltalMoney());
+            }
+            if (vO.getIdGroupTable() != null) {
+                GroupTable groupTable = groupTableRepository.findById(vO.getIdGroupTable()).orElseThrow(() -> new NoSuchElementException());
+                bean.setGroupTable(groupTable);
+            }
+            if (vO.getIdEmployee() != null) {
+                Employee employee = employeeRepository.findById(vO.getIdEmployee()).orElseThrow(() -> new NoSuchElementException());
+                bean.setEmployee(employee);
+            }
+            invoiceRepository.save(bean);
+        }
+        else {
+            throw new NoSuchElementException();
+        }
     }
 
     public List<InvoiceNoMapPing> findByIdGroupTableAndStatus(String id, String status) {
